final_project/requirements.txt
python-docx
re
pytest